+++ contracts/mocks/DCAPair/DCAPairSwapHandler.sol
@@ -6,7 +6,7 @@ import '../../DCAPair/DCAPairSwapHandler.sol';
 import './DCAPairParameters.sol';
 
 contract DCAPairSwapHandlerMock is DCAPairSwapHandler, DCAPairParametersMock {
-  uint256 private _customTimestamp;
+  uint32 private _customTimestamp;
 
   constructor(
     IERC20Detailed _token0,
@@ -33,11 +33,11 @@ contract DCAPairSwapHandlerMock is DCAPairSwapHandler, DCAPairParametersMock {
     return _getAmountToSwap(_tokenAddress, _swap);
   }
 
-  function setBlockTimestamp(uint256 _blockTimestamp) public {
+  function setBlockTimestamp(uint32 _blockTimestamp) public {
     _customTimestamp = _blockTimestamp;
   }
 
-  function _getTimestamp() internal view override returns (uint256 _blockTimestamp) {
+  function _getTimestamp() internal view override returns (uint32 _blockTimestamp) {
     _blockTimestamp = (_customTimestamp > 0) ? _customTimestamp : super._getTimestamp();
   }
 
@@ -55,7 +55,7 @@ contract DCAPairSwapHandlerMock is DCAPairSwapHandler, DCAPairParametersMock {
     swapAmountAccumulator[_tokenAddress] = _swapAmountAccumulator;
   }
 
-  function setLastSwapPerformed(uint256 _lastSwapPerformend) public {
+  function setLastSwapPerformed(uint32 _lastSwapPerformend) public {
     lastSwapPerformed = _lastSwapPerformend;
   }
 }
